Rename package secret to endpoint.
